+++ Razor/src/Razor/Systems/RSMaterialPass.cpp
@@ -25,7 +25,7 @@ namespace Razor
 			}
 			ShaderProperty& Property = data.mEntityRenderProperties.Properties[RenderingEntity];
 			Property.GeneratePropertySlots(EntityMat.Materials.size());
-			for (int i = 0; i < EntityMat.Materials.size(); i++)
+			for (size_t i = 0; i < EntityMat.Materials.size(); i++)
 			{
 				ShaderPropertySlot& Slot = Property.GetPropertySlot(i);
 				Slot.AddProperty<int>("material.texture_diffuse", 0);
